Scope global file upload to file library
Fixes #399, fixes #262
